% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_output.R
\name{pprint}
\alias{pprint}
\title{Selects and returns a descriptive value from the LM statistic table}
\usage{
pprint(statt, sel = NA, table = 1, pretty = 1, ...)
}
\arguments{
\item{statt}{\emph{stat}istic \emph{t}able of LM statistics generated by \code{StartPLMScoRe}}

\item{sel}{list of selection criteria (see details)}

\item{table}{returns results as a table with (default), \code{table=0} returns only the values}

\item{pretty}{rounds numeric output to two decimal points (default)}

\item{...}{further arguments (currently not implemented)}
}
\value{
the selected descriptive statistics
}
\description{
Selects and returns a descriptive value from the LM statistic table
}
\details{
The function \code{StartPLMScoRe()} generates a large table with LM statistics,
currently 318 different statistics. \code{pprint} can be used to select and return
selected values from this table.
The input \code{stats} refers to the LM statistic table. If you have run the command
\code{outputXY<-StartPLMScoRe()} this table can be found with \code{outputXY$Stats} or
\code{outputXY[[3]]}.
With the input \code{sel} you specify your selection criteria, for example \code{sel=
c("PLMnr", "no./hour", "TST")} will return the PLMnr index for total sleep time. The
selection operates on three data types:

\describe{
  \item{LM type}{\describe{
  \item{Main category LM:}{CLM, nonCLM, PLM}
 \item{ Modifier 1:}{l (left), r (right), b (bilateral), m (monolateral).
  These modifiers are added before the main category with and underscore separating the two,
  eg. r_CLM, b_nonCLM}
  \item{Modifier 2:}{r (respiratory event associated), nr (non respiratory event associated), a
  (arousal associated).
  These modifies are added after the main category, if there are more than two these are
  separated by and underscore, eg. PLMnr, PLMa, but CLMnr_a.}
  \item{Other categories:}{IMI, IMInr, Arousal, R events, Sleep/wake}}}

  \item{Statistic}{\describe{
  \item{for LM:}{number, no./hour, duration; mean, duration; SD, duration; min,
  duration; max}
  \item{for IMI:}{log; mean, log;SD, <10s; number, >90s; number, 10-90s; number}
  \item{for Sleep/wake:}{duration}
  \item{for Arousal:}{number, no./hour}
  \item{for R events:}{number, no./hour, \% with CLM }}}

  \item{Sleep/wake stage}{TIB, TST, N1, N2, N3, REM, NREM}
  }
Include in the \code{sel=c()} input all items you want to retrieve from the LM statistic table.
If you do not specify anything for any of the three categories (LM type, Statistic, Sleep/wake)
all instances will be selected:

sel=c("PLMnr", "no./hour") will return PLMnr indices for TIB, TST, Wake, N1, N2, N3, REM, NREM

sel=c("IMI") will return all available statistics for all sleep/wake stages

You can also specify multiple selections from a single category:

sel=c("CLM", "CLMnr", "number", "N1", "REM") will return the number of CLM and CLMnr during N1 and
REM sleep.
}
\examples{
# Return PLMS index from LM statistics table \emphasis{lmstat}
\dontrun{pprint(lmstat, sel=c("PLM", "no./hour", "TST"))}

# Return PLMS index from LM statistics table \emphasis{lmstat} but do not round the result
\dontrun{pprint(lmstat, sel=c("PLM", "no./hour", "TST"), pretty=0)}

#Return PLMS and PLMSnr index
\dontrun{pprint(lmstat, sel=c("PLM", "PLMnr", "no./hour", "TST"))}

}
