% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PLMSCORE-EDF.R
\name{StartPLMScoRe_edf}
\alias{StartPLMScoRe_edf}
\title{PLMScoRe function for edf plus annotation files}
\usage{
StartPLMScoRe_edf(fn = NA, rrules = NA, output = NA, ...)
}
\arguments{
\item{fn}{Filename and path of edf+ annotation file}

\item{rrules}{Chosen definition or respiratory event related leg movements (1 = -2 to 10.25, 2 = -0.5 to 0.5)}

\item{output}{chosen output options (screen, csv, pdf)}

\item{...}{Additional arguments (currently not specified)}
}
\value{
a PLMScoRe object
}
\description{
Will read in an edf+ annotation file and return descriptors of
leg movements
}
\details{
If you are using \code{StartPLMScore_edf} with no input arguments (\code{resultXY<-StartPLMScore_edf()}),
you will be asked to chose the definition of respiratory event related leg movements
and your output options. Alternatively (see examples below) you can specify both directly.
}
\examples{
#Will run the function interactively

\dontrun{plmresultsXY<-StartPLMScoRe_edf()}

#Will run the function automatically without further user input

\dontrun{plmresultsxy<-StartPLMScore_edf(fn="D:\\\\Test\\\\XY-Events.edf", rrules=1, output=c("screen", "csv", "pdf"))}

#Example of batch processing

#You want to make sure that you select only edf files
#This will generate a list of all edf files in that directory

\dontrun{fnlist<-list.files("D:\\\\Test\\\\Scorings\\\\", pattern=".edf", full.names=TRUE)}

\dontrun{for(i in 1:length(fnlist)) StartPLMScore_edf(fn=fnlist[i], rrules=1, output=c("screen", "csv", "pdf"))}
}

